Renamed the hand "Two Pairs" to "Two Pair".
Nitpick, but the name of the hand is "Two Pair".
https://en.wikipedia.org/wiki/Two_pair#Two_pair

English is a funny language.

This commit does not correct the name of the hand in any images (for
example data/gfx/gui/table/default/handranking.png) because the
corresponding GIMP templates do not have separate text layers, making
them difficult to edit.